+++ requests/core.py
@@ -12,12 +12,13 @@ This module implements the main Requests system.
 """
 
 __title__ = 'requests'
-__version__ = '0.3.4'
-__build__ = 0x000304
+__version__ = '0.4.0'
+__build__ = 0x000400
 __author__ = 'Kenneth Reitz'
 __license__ = 'ISC'
 __copyright__ = 'Copyright 2011 Kenneth Reitz'
 
 
-from .models import HTTPError, auth_manager
-from .api import *
+
+from models import HTTPError, auth_manager
+from api import *