+++ HISTORY.rst
@@ -1,8 +1,16 @@
 History
 -------
 
-0.3.4
-+++++
+0.4.0 (2011-05-15)
+++++++++++++++++++
+
+* Response.history: list of redirected responses
+* Case-Insensitive Header Dictionaries!
+* Unicode URLs
+
+
+0.3.4 (2011-05-14)
+++++++++++++++++++
 
 * Urllib2 HTTPAuthentication Recursion fix (Basic/Digest)
 * Internal Refactor